requirements.txt
streamlit
pandas
requests
google-genai
httpx
torchvision
scikit-learn